sheets-to-json/requirements.txt
numpy==2.3.2
pandas==2.3.2
python-dotenv==1.1.1
requests==2.32.5

